backend/phishing-function/requirements.txt
appwrite
rapidfuzz
requests
google-generativeai
